Regex positive and negative matching
Added two extra command line parameters to projectwatch.py
(--filename, --notlike). Both are passed down to the command
if a raw command is not given. For both a regex is expected.
For the --filename the regex must match or the event is ignored.
For the --notlike the regex must NOT match or the event is
ignored.

The parameters are both passed to clsinc (via --match/--nmatch)
and the files that are parsed for classes/interfaces must match
and not match the regexes accordingly or else they are ignored.